django/requirements-async.txt
channels_rabbitmq==4.0.0
channels_redis==3.31
daphne==3.0.2
